+++ serpentTools/tests/test_branching.py
@@ -23,7 +23,7 @@ class _BranchTesterHelper(unittest.TestCase):
         cls.expectedBranches = {('nom', 'nom', 'nom')}
         cls.expectedUniverses = {
             # universe id, burnup, step
-            (0, 0, 0),
+            ("0", 0, 0),
         }
         with rc:
             rc['serpentVersion'] = '2.1.29'
@@ -78,9 +78,15 @@ class BranchContainerTester(_BranchTesterHelper):
     def setUpClass(cls):
         _BranchTesterHelper.setUpClass()
         cls.refBranchID = ('nom', 'nom', 'nom')
-        cls.refUnivKey = (0, 0, 0)
+        cls.refUnivKey = ("0", 0, 0)
         cls.refBranch = cls.reader.branches[cls.refBranchID]
         cls.refUniv = cls.refBranch.universes[cls.refUnivKey]
+
+    def test_universeIDWorkaround(self):
+        """Test that, for now, integer universe ids work"""
+        # Remove this test for versions >= 0.8.0
+        key = (int(self.refUnivKey[0]), ) + self.refUnivKey[1:]
+        self.assertIs(self.refUniv, self.refBranch.universes[key])
 
     def test_loadedUniv(self):
         """Verify the reference universe has the correct data loaded"""
@@ -149,7 +155,7 @@ class BranchWithUncsTester(unittest.TestCase):
             ],
         },
     }
-    BRANCH_UNIVKEY = (0, 0, 0)
+    BRANCH_UNIVKEY = ("0", 0, 0)
 
     def setUp(self):
         fp = getFile('demo_uncs.coe')